backend_requirements.txt
# Kutubxona Boshqaruv Tizimi - Python Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Environment Variables
python-dotenv==1.0.0

# Validation
pydantic[email]==2.5.0
email-validator==2.1.0

# CORS
python-multipart==0.0.6

# Optional: Email Notifications
# aiosmtplib==3.0.1

# Optional: Background Tasks
# celery==5.3.4
# redis==5.0.1

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality (Optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1